requirements.txt
fastapi==0.128.0
openai==2.15.0
pydantic==2.12.5
pymongo==4.16.0
python-dotenv==1.2.1
jinja2
